+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Removed image related fields from schema.
+  [iham]
 
 
 1.0.0 (2019-05-10)